feat(us1): Complete Blazor UI components and service registration (T041-T045)
Implemented complete Blazor Server UI for User Story 1 (CREADOR report management):

**Reusable Components** (T043, T044):
- CascadingSelect.razor: Generic cascading dropdown with loading states
- SuggestionInput.razor: Autocomplete input with debounced filtering

**Page Components** (T041, T042):
- ReportForm.razor: Complete Type A report form with 6 sections:
  - Datos de los Hechos (datetime)
  - Datos de la Víctima (sexo, edad, flags)
  - Ubicación Geográfica (zona → sector → cuadrante cascade)
  - Tipo de Incidencia (delito, tipo de atención with suggestions)
  - Detalles Operativos (turno, tipo de acción, traslados)
  - Narrativa (hechos reportados, acciones realizadas, observaciones)
- ReportList.razor: Report history with filtering, pagination, responsive design

**Service Registration** (T045):
- Added IReportService → ReportService
- Added ICatalogService → CatalogService (Infrastructure layer)
- Added IReportRepository → ReportRepository

**Features**:
- Two-way data binding with custom EventCallback handlers
- Real-time cascading dropdown updates (zona → sector → cuadrante)
- Autocomplete suggestions with debounced search
- Draft save and submit workflows
- Authorization: CREADOR role required
- Mobile-responsive layouts (table for desktop, cards for mobile)
- Validation messages with Bootstrap styling
- Loading states for async operations

**Status**: User Story 1 implementation complete (T021-T046). Production code builds successfully. Test project package dependencies pending (Moq, Microsoft.AspNetCore.Mvc.Testing).

+++ tests/Ceiba.Core.Tests/ReporteIncidenciaTests.cs
@@ -208,7 +208,7 @@ public class ReporteIncidenciaTests
         // Arrange
         var report = new ReporteIncidencia
         {
-            TipoDeAccion = tipoDeAccion
+            TipoDeAccion = (short)tipoDeAccion
         };
 
         // Act
@@ -227,7 +227,7 @@ public class ReporteIncidenciaTests
         // Arrange
         var report = new ReporteIncidencia
         {
-            TipoDeAccion = tipoDeAccion
+            TipoDeAccion = (short)tipoDeAccion
         };
 
         // Act
@@ -246,7 +246,7 @@ public class ReporteIncidenciaTests
         // Arrange
         var report = new ReporteIncidencia
         {
-            Traslados = traslados
+            Traslados = (short)traslados
         };
 
         // Act
@@ -265,7 +265,7 @@ public class ReporteIncidenciaTests
         // Arrange
         var report = new ReporteIncidencia
         {
-            Traslados = traslados
+            Traslados = (short)traslados
         };
 
         // Act
